elimino client y gateway anteriores
